split database files into functions and build files
