+++ package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node app.js",
     "dev": "./node_modules/.bin/nodemon app.js",
-    "lint": "./node_modules/.bin/eslint 'app.js'"
+    "lint": "./node_modules/.bin/eslint"
   },
   "repository": {
     "type": "git",
